integrated requetsed changes in products api and added file upload functionality
